% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_genes.R
\docType{data}
\name{bcr.string}
\alias{bcr.string}
\title{Regular expression pattern for BCR genes}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcr.string
}
\description{
Pattern to identify B-cell receptor genes
}
\keyword{datasets}
