% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_conversion.R
\name{convert_genes}
\alias{convert_genes}
\title{Convert genes between species efficiently}
\usage{
convert_genes(genes, org.from = "human", org.to = "mouse", one.to.many = FALSE)
}
\arguments{
\item{genes}{Vector of gene symbols to convert}

\item{org.from}{Source organism ("human" or "mouse")}

\item{org.to}{Target organism ("human" or "mouse")}

\item{one.to.many}{If TRUE, returns all possible mappings}
}
\value{
Vector of converted gene symbols
}
\description{
Convert genes between species efficiently
}
