% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conversion.R
\name{convert_orthologs_df}
\alias{convert_orthologs_df}
\title{Convert orthologs between human and mouse for a data frame}
\usage{
convert_orthologs_df(df, gene_column, mode = "human_to_mouse")
}
\arguments{
\item{df}{A data frame of gene symbols to convert}

\item{mode}{The direction of conversion, either "human_to_mouse" or "mouse_to_human"}

\item{many.to.one}{Logical. Default is FALSE to return only unique mappings (one/many to one). If TRUE, returns all possible mouse gene mappings (one to many), including cases where one human gene maps to multiple mouse genes.}
}
\description{
Converts a data frame of gene symbols between human and mouse species.
}
