% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conversion.R
\name{convert_orthologs_matrix}
\alias{convert_orthologs_matrix}
\title{Convert orthologs between human and mouse for an gene expression matrix}
\usage{
convert_orthologs_matrix(matrix, mode = "human_to_mouse", many.to.one = TRUE)
}
\arguments{
\item{matrix}{A matrix of gene symbols to convert}

\item{mode}{The direction of conversion, either "human_to_mouse" or "mouse_to_human"}

\item{many.to.one}{Logical. Default is FALSE to return only unique mappings (one/many to one). If TRUE, returns all possible mouse gene mappings (one to many), including cases where one human gene maps to multiple mouse genes.}
}
\description{
Converts a matrix of gene symbols between human and mouse species.
}
