% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conversion.R
\name{convert_orthologs_vector}
\alias{convert_orthologs_vector}
\title{Convert orthologs between human and mouse for a vector of genes}
\usage{
convert_orthologs_vector(genes, mode = "human_to_mouse", one.to.many = TRUE)
}
\arguments{
\item{genes}{A vector of gene symbols to convert}

\item{mode}{The direction of conversion, either "human_to_mouse" or "mouse_to_human"}

\item{one.to.many}{Logical. Default is FALSE to return only unique mappings (one/many to one). If TRUE, returns all possible mouse gene mappings (one to many), including cases where one human gene maps to multiple mouse genes.}
}
\value{
A vector of converted gene symbols
}
\description{
Converts a vector of gene symbols between human and mouse species.
}
\examples{
\dontrun{
mouse_genes <- c("Trp53", "Cd4", "Cd8a")
human_genes <- convert_orthologs_vector(mouse_genes, mode = "mouse_to_human", one.to.many = TRUE)
print(human_genes)
}
}
