% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{df_to_list}
\alias{df_to_list}
\title{Convert Data Frame to List}
\usage{
df_to_list(df)
}
\arguments{
\item{df}{A data frame to convert}
}
\value{
A list of vectors where each element corresponds to a column from the input data frame.
}
\description{
Converts a data frame into a list of vectors, removing empty strings.
}
\examples{
\dontrun{
my_df <- data.frame(c("A", "B", "C"), c("D", "E"), c("F", "G", "H", "I"))
result <- df_to_list(my_df)
print(result)
}
}
