% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Package startup message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{The library where the package is installed}

\item{pkgname}{The name of the package}
}
\description{
Displays the package version when the package is attached
}
\keyword{internal}
