% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Initialize package when loaded}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The library where the package is installed}

\item{pkgname}{The name of the package}
}
\description{
This function is called when the package is loaded via library() or require()
}
\keyword{internal}
