% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.strpip_env}
\alias{.strpip_env}
\title{Internal Package Environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.strpip_env
}
\description{
This environment is used to store package-level data that needs to persist
between function calls without polluting the user's global environment.
}
\keyword{internal}
