% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_functions.R
\name{find_annotation_file}
\alias{find_annotation_file}
\title{Find OmniPath Annotation File}
\usage{
find_annotation_file(org, release)
}
\arguments{
\item{org}{Organism type ("human" or "mouse")}

\item{release}{Ensembl release version}
}
\value{
Path to the annotation file
}
\description{
Internal function to locate the correct annotation file based on organism and release.
}
\keyword{internal}
