% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_genes.R
\docType{data}
\name{hb.string}
\alias{hb.string}
\title{Regular expression pattern for hemoglobin genes}
\format{
An object of class \code{character} of length 1.
}
\usage{
hb.string
}
\description{
Pattern to identify hemoglobin genes
}
\keyword{datasets}
