% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conversion.R
\name{import_biomart_local}
\alias{import_biomart_local}
\title{Import BioMart data from local file}
\usage{
import_biomart_local(release = "105")
}
\arguments{
\item{release}{Ensembl release number. Default is "105".}
}
\value{
A data frame containing gene mappings between species
}
\description{
Retrieves gene mapping data from a local TSV file instead of querying Ensembl.
This is faster than \code{import_biomart()} but requires the data file to be present.
}
\examples{
\dontrun{
biomart_dict <- import_biomart_local()
head(biomart_dict)
}
}
