% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_detect.R
\docType{data}
\name{mhc.string}
\alias{mhc.string}
\title{Regular expression pattern for MHC genes}
\format{
An object of class \code{character} of length 1.
}
\usage{
mhc.string
}
\description{
Pattern to identify major histocompatibility complex genes
}
\keyword{datasets}
