% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_detect.R
\docType{data}
\name{mt.string}
\alias{mt.string}
\title{Regular expression pattern for mitochondrial genes}
\format{
An object of class \code{character} of length 1.
}
\usage{
mt.string
}
\description{
Pattern to identify mitochondrial genes
}
\keyword{datasets}
