% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_detect.R
\docType{data}
\name{rb.string}
\alias{rb.string}
\title{Regular expression pattern for ribosomal genes}
\format{
An object of class \code{character} of length 1.
}
\usage{
rb.string
}
\description{
Pattern to identify ribosomal protein genes
}
\keyword{datasets}
