% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_detect.R
\docType{data}
\name{tcr.string}
\alias{tcr.string}
\title{Regular expression pattern for TCR genes}
\format{
An object of class \code{character} of length 1.
}
\usage{
tcr.string
}
\description{
Pattern to identify T-cell receptor genes
}
\keyword{datasets}
