% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{validate_gene_symbols}
\alias{validate_gene_symbols}
\title{Check if Gene Symbols Are Valid}
\usage{
validate_gene_symbols(genes, org = "human")
}
\arguments{
\item{genes}{Character vector of gene symbols to check}

\item{org}{Organism type, either "human" or "mouse"}
}
\value{
Logical indicating whether the genes are valid
}
\description{
Internal function to check if gene symbols are valid for a given organism.
}
\keyword{internal}
