% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_org}
\alias{validate_org}
\title{Validate Org Parameters}
\usage{
validate_org(org, genes = NULL)
}
\arguments{
\item{org}{Organism type, either "human" or "mouse"}

\item{genes}{Gene symbols to validate}
}
\value{
Logical indicating whether the genes are valid
}
\description{
Validate Org Parameters
}
\keyword{internal}
