% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-biomart.R, R/utils.R
\name{validate_organism}
\alias{validate_organism}
\title{Validate Organism Parameter}
\usage{
validate_organism(org)

validate_organism(org)
}
\arguments{
\item{org}{Organism string to validate}
}
\value{
The validated organism or an error if invalid

A validated organism string ("human" or "mouse")
}
\description{
Internal function to validate the organism parameter.

Internal function to validate the organism parameter
}
\keyword{internal}
