% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{write_gct}
\alias{write_gct}
\title{Write GCT File}
\usage{
write_gct(df, file)
}
\arguments{
\item{df}{Data frame to save}

\item{save_name}{Name of the output file}

\item{save_dir}{Directory to save the file in. Defaults to current working directory.}
}
\value{
No return value. Creates a GCT file at the specified location.
}
\description{
Saves a data frame as a Gene Cluster Text (GCT) file.
The output file will be formatted according to the GCT specification.
}
\examples{
\dontrun{
write_gct(expression_df, "my_expression.gct", "output/")
}
}
