% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{validate_organism}
\alias{validate_organism}
\title{Validate Organism Parameter}
\usage{
validate_organism(org)
}
\arguments{
\item{org}{Organism to validate}
}
\value{
The validated organism or an error if invalid
}
\description{
Internal function to validate the organism parameter.
}
\keyword{internal}
