% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_gmt}
\alias{write_gmt}
\title{Write GMT File}
\usage{
write_gmt(df, save_name = "", save_dir = getwd())
}
\arguments{
\item{df}{Data frame to save}

\item{save_name}{Name of the output file}

\item{save_dir}{Directory to save the file in. Defaults to current working directory.}
}
\value{
No return value. Creates a GMT file at the specified location.
}
\description{
Saves a data frame as a Gene Matrix Transposed (GMT) file.
The output file will be formatted according to the GMT specification.
}
\examples{
\dontrun{
write_gmt(gene_sets_df, "my_genesets.gmt", "output/")
}
}
